+++ src/init.c
@@ -132,6 +132,7 @@ extern SEXP vctrs_normalize_encoding(SEXP);
 extern SEXP vctrs_order(SEXP, SEXP, SEXP);
 extern SEXP vctrs_order_locs(SEXP, SEXP, SEXP);
 extern SEXP vctrs_unrep(SEXP);
+extern SEXP vctrs_fill_missing(SEXP, SEXP, SEXP);
 
 
 // Maturing
@@ -281,6 +282,7 @@ static const R_CallMethodDef CallEntries[] = {
   {"vctrs_order",                      (DL_FUNC) &vctrs_order, 3},
   {"vctrs_order_locs",                 (DL_FUNC) &vctrs_order_locs, 3},
   {"vctrs_unrep",                      (DL_FUNC) &vctrs_unrep, 1},
+  {"vctrs_fill_missing",               (DL_FUNC) &vctrs_fill_missing, 3},
   {NULL, NULL, 0}
 };